fsck: verify commit graph when implicitly enabled
Change fsck to check the "core_commit_graph" variable set in
"repo-settings.c" instead of reading the "core.commitGraph" variable,
this fixes a bug where we wouldn't start "commit-graph verify" if the
config key was missing, which since 31b1de6a09 (commit-graph: turn on
commit-graph by default, 2019-08-13) has meant that it's on by
default.

Add tests to "t5318-commit-graph.sh" to check that this works, using
the detection of a corrupt commit-graph as a method of seeing if
"commit-graph verify" was invoked.

Signed-off-by: Glen Choo <chooglen@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>