revisions API users: add release_revisions() where "prune_data" is free'd
Add a release_revisions() to various users of "struct rev_list" that
reach into the "struct rev_info" and clear the "prune_data"
already. In a subsequent commit we'll teach release_revisions() to
clear this itself, but in the meantime let's invoke
release_revisions() here to clear anything else we may have missed,
and for reasons of having consistent boilerplate.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>