cache-tree API: make write_{cache,index}_as_tree() return an enum
Change the return value of write_{cache,index}_as_tree() to be an
enum. This allows us to assert that
e.g. write_index_as_tree_internal() will never return other values,
and furthermore that it won't return WRITE_TREE_UNREADABLE_INDEX.

In a subsequent commit callers upstream of this will be able to do the
same to simplify and improve their error handling.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>