stash: don't assume why write_cache_as_tree() fails
Improve the messages emitted by "git stash apply" when
write_cache_as_tree() fails to say that it failed, but not to assume
that we're in a merge. That's only on of its failure states.

This code should be improved to hold a lock throughput to eliminate
this edge case, i.e. we'll call refresh_and_write_cache() earlier in
this function which locks the index, but then we'll release the lock,
only to acquire it again with hold_lock_file_for_update() in
write_index_as_tree(). a "git stash apply" on my box will acquire the
"index.lock" 4 times to do its work, and then once again as it
executes "git status". #leftoverbits

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>