builtins: normalize subcommand patterns with usage_msg_optf()
Now that we have a usage_msg_optf() we can migrate commit-graph,
multi-pack-index etc. to a better pattern of always telling the user
what's wrong with the command-line, instead of only spewing the (often
large) usage message at them.

We'll now say if no subcommand is given, and if the subcommand is
unknown. The changed commands were in various states of getting that
right.

Let's also change the wording to be the same for all these error
messages, so that our translators don't have to translate multiple
variants of the same thing.

This changes the behavior from emitting a "error: " to emitting a
"fatal: ", but that's OK, we'll emit the same exit code as
before (129). One test was being overly specific and needed to be
adjusted for this.

For the code in "bundle.c" the "result ? 1 : 0" was already redundant
per my reading of the code, but let's retain it in the new do_cmd()
helper function with a shorter !! idiom.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>