[TODO] bundle-uri client: support incremental "fetch"
Support incremental "fetch" when downloading bundles.

This can be tested with the recipe noted in a preceding commit to
create the "git-master-only.bdl" and "git-master-to-next.bdl"
files.

First let's clone "master" vi a bundle.

Note that for both requests we supply bundles for both "master" and
"next", but even then discover for the first request that we only
needed the one for "master". For the second request we then discover
that we have the OIDs from the "master" bundle, but not the "next"
one, and will use it to incrementally fetch:

    rm -rf /tmp/git.git &&
    ./git \
	-c protocol.version=2 \
        -c fetch.uriProtocols=file \
        -c transfer.injectBundleURI="file:///tmp/git-master-only.bdl" \
        -c transfer.injectBundleURI="file:///tmp/git-master-to-next.bdl" \
	clone --bare --no-tags --single-branch --branch master --template= \
	--verbose --verbose \
	https://github.com/git/git.git /tmp/git.git &&
    ./git \
	-C /tmp/git.git \
	-c protocol.version=2 \
        -c fetch.uriProtocols=file \
        -c transfer.injectBundleURI="file:///tmp/git-master-only.bdl" \
        -c transfer.injectBundleURI="file:///tmp/git-master-to-next.bdl" \
	fetch \
	--verbose --verbose \
	origin 'refs/heads/seen:refs/heads/seen'

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>