serve.c: split up "advertise" from "value"
Simplify the "struct protocol_capability" by untangling whether
something is advertised from whether it would like to add a value to
the capability.

All the capabilities except one wanted to be advertised
unconditionally, so by saying that they're advertised by default we
can get rid of that callback for all but one of them, as well as the
"always_advertise" callback boilerplate added in 72d0ea0056b (ls-refs:
introduce ls-refs server command, 2018-03-15).

What those previous "advertise" functions mostly wanted to do was not
to optionally hide the capability, but to optionally add a capability
value. This is now done via a new "value" callback. This also makes
the previous flow of the "advertise" callback simpler, as we'd either
call it with a "struct strbuf *" or "NULL" depending on whether or not
we were calling it from protocol_v2_advertise_capabilities().

The code changes to "upload-pack.c" mostly go away under "-w",
i.e. we're just getting rid of the "if (value) {" and "return 1",
since the remaining function will only be called to optionally append
to the capability value.

To reduce the inline verbosity of checking if c->advertise is NULL or
not add a simple static "is_advertised" wrapper function in serve.c.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>