serve tests: add missing test for omitted object-format
As c7d3aabd279 (serve: provide "receive" function for object-format
capability, 2021-09-14) correctly points out we need to error out if
we have a SHA-256 client talking to a SHA-1 server in a case where
that client didn't send an "object-format" line.

But if the check in process_request() amended in that commit was moved
to object_format_receive() all our tests would pass, since we didn't
test for this edge case, let's do so.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>