serve tests: test behavior of "session-id"
There are tests for how "session-id" works added in
23bf486acae (transport: log received server session ID, 2020-11-11),
but those are not the light "test-tool serve-v2" kind, they run a full
v2 server and test the behavior of "git-upload-pack".

Let's add basic tests for its behavior here, to make it easier to
change serve.c itself, and document a bug in its implementation. It
should error if no arguments are given, as "object-format" does (see
c7d3aabd279 (serve: provide "receive" function for object-format
capability, 2021-09-14)), but it doesn't.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>