serve: simplify command and capability parsing
As noted in 108c265f272 (serve: reject bogus v2 "command=ls-refs=foo",
2021-09-15) we'll call get_capability() twice. This change untangles
the request parsing so that we'll only do so once.

Similarly for is_advertised() we'll now only call it twice (after
calling it once in "protocol_v2_advertise_capabilities()"), before
we'd check if something was a command, then loop over the known
capabilities, and if it wasn't a command we'd loop over them again. By
simply stripping "command=" away at the start we can do away with all
of this and other duplicate and redundant checks.

The error messaging being added here is deliberately bad, so as to be
bug-for-bug compatible with the existing behavior and tests, but it
will be improved in subsequent commits.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>