serve: make specifying .command and .receive a BUG()
It isn't against the protocol v2 spec to have a command that would
like to have a "receive" callback, i.e. a hypothetical "food" which a
client would request as:

    food=spicy
    command=food
    0000

Although doing so would not have any obvious benefit over merely
specifying that pseudo-argument as an argument to the command itself,
as all current commands that require that sort of thing do:

    command=food
    0001
    spicy
    0000

See the documentation added in ed10cb952d3 (serve: introduce
git-serve, 2018-03-15) for reference.

But let's make doing so a BUG(), since our implementation definitely
does not expect this state, we can always remove this BUG() if we ever
add such an odd command in the future.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>