serve: specify invalid v.s. unknown in error output
Change the serve.c error output to clearly indicate whether something
is a known command, unknown, or known but not advertised by the
server.

This leaks more information than we did before, but I think in this
case it's a non-issue. It's hardly a mystery to anyone who'd be trying
to probe a v2 server using git.git's git-upload-pack why
e.g. "session-id" would be supported, or not.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>