serve: error if "session-id" is not given an argument
Fix a looseness in how "session-id" and other future "receive"
capabilities are handled. In c7d3aabd279 (serve: provide "receive"
function for object-format capability, 2021-09-14) we started emitting
errors if "object-format" was not provided with an argument, but the
similar "session-id" argument was not changed.

Let's bring this in line with the spec, which indicates that these
capabilities MUST get an argument.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>