serve: s/invalid/unknown/g in "command" error output
Bring the error output when unknown commands are requested in line
with the existing "unknown capability" error output.

It's possible that a client requested an unknown command, i.e. one
whose "key" is not (per "protocol-v2.txt") 1*(ALPHA | DIGIT | "-_").

But more than likely the client just requested an unknown command, we
could additionally parse the line to see if the key is invalid, but
let's not bother.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>