Makefile: don't set up "perl/build" rules under NO_PERL=Y
Do not define LIB_{PERL,CPAN}{,_GEN} if NO_PERL is defined. This
changes no functionality, but makes it clear which of these rules are
needed under NO_PERL=Y. See 20d2a30f8ff (Makefile: replace
perl/Makefile.PL with simple make rules, 2017-12-10) for the initial
implementation.

We do for better or worse rely on "install-doc" calling
"install-man-perl" regardless of whether NO_PERL=Y is defined or
not. Let's add a comment about that adjacent to the rules that build
perl/build.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>